BACKEND_OLD/backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.3.2
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
python-dotenv==1.1.1
flask-cors==4.0.0
gunicorn==21.2.0
xgboost==1.7.6
lightgbm==4.1.0
folium==0.15.1
requests==2.32.5
gtfs-realtime-bindings==1.0.0
protobuf==4.25.0
